refactor Repay flow to use Formik hook
